requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
Pillow==10.4.0
onnxruntime==1.18.1
pytest==8.2.2
black==24.4.2
ruff==0.4.8
